pyproject.toml
[project]
name = "automate"
version = "0.1.0"
description = "Privacy-first email & content automation with local LLM support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.159.0",
    "anthropic>=0.42.0",
    "google-generativeai>=0.8.0",
    "beautifulsoup4>=4.12.3",
    "markdownify>=0.14.1",
    "python-dateutil>=2.9.0",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services", "config"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long — formatter handles this
    "B008",   # do not perform function call in argument defaults (FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["services", "config"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
